ISPN-1278 Remove unused test class and useless test package.
