+++ cdi/src/test/java/org/infinispan/cdi/test/interceptor/CacheRemoveAllInterceptorTest.java
@@ -25,8 +25,8 @@ package org.infinispan.cdi.test.interceptor;
 import org.infinispan.Cache;
 import org.infinispan.cdi.test.interceptor.service.CacheRemoveService;
 import org.infinispan.cdi.test.interceptor.service.Custom;
+import org.infinispan.cdi.test.interceptor.service.CustomCacheKeyGenerator;
 import org.infinispan.cdi.test.interceptor.service.GreetingService;
-import org.infinispan.cdi.test.interceptor.service.generator.CustomCacheKeyGenerator;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.testng.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
