+++ cdi/src/test/java/org/infinispan/cdi/test/interceptor/service/CacheRemoveService.java
@@ -22,8 +22,6 @@
  */
 package org.infinispan.cdi.test.interceptor.service;
 
-import org.infinispan.cdi.test.interceptor.service.generator.CustomCacheKeyGenerator;
-
 import javax.cache.interceptor.CacheRemoveAll;
 import javax.cache.interceptor.CacheRemoveEntry;
